light and bit packing

+++ Benchmark/BenchMarks/ChunkToSuperChunkBenchmark.cs
@@ -10,7 +10,7 @@ public class ChunkToSuperChunkBenchmark
     [Benchmark]
     public unsafe void testSuperChunkMethod() {
         BlockData[,,] blocks = new BlockData[16,16,16];
-        int index = 0;
+        short index = 0;
         for (int i = 0; i < 16; i++) {
             for (int j = 0; j < 16; j++) {
                 for (int k = 0; k < 16; k++) {
@@ -39,7 +39,7 @@ public class ChunkToSuperChunkBenchmark
     [Benchmark]
     public unsafe void testSuperChunkMethodParralle() {
         BlockData[,,] blocks = new BlockData[16,16,16];
-        int index = 0;
+        short index = 0;
         for (int i = 0; i < 16; i++) {
             for (int j = 0; j < 16; j++) {
                 for (int k = 0; k < 16; k++) {
@@ -67,7 +67,7 @@ public class ChunkToSuperChunkBenchmark
     [Benchmark]
     public unsafe void testSuperChunkMethodFor() {
         BlockData[,,] blocks = new BlockData[16,16,16];
-        int index = 0;
+        short index = 0;
         for (int i = 0; i < 16; i++) {
             for (int j = 0; j < 16; j++) {
                 for (int k = 0; k < 16; k++) {
